requirements.txt
streamlit
pandas
wordcloud
matplotlib
mecab-python3
ipadic
